conversion from 'size_t' to 'int'

+++ src/ltm/bn_mp_export.c
@@ -65,7 +65,7 @@ int mp_export(void* rop, size_t* countp, int order, size_t size,
 
 			*byte = (unsigned char)((j == ((size - nail_bytes) - 1)) ? (t.dp[0] & odd_nail_mask) : (t.dp[0] & 0xFF));
 
-			if ((result = mp_div_2d(&t, ((j == ((size - nail_bytes) - 1)) ? (8 - odd_nails) : 8), &t, NULL)) != MP_OKAY) {
+			if ((result = mp_div_2d(&t, (int)((j == ((size - nail_bytes) - 1)) ? (8 - odd_nails) : 8), &t, NULL)) != MP_OKAY) {
 				mp_clear(&t);
 				return result;
 			}
